hotfix: fix roles and permissions management

+++ app/Http/Controllers/ProductController.php
@@ -6,6 +6,7 @@ use App\Traits\Responses;
 use Illuminate\Http\Request;
 use App\Http\Requests\ProductRequest;
 use App\Repositories\ProductRepository;
+use Illuminate\Support\Facades\Gate;
 
 class ProductController extends Controller
 {
@@ -18,6 +19,7 @@ class ProductController extends Controller
      */
     public function index()
     {
+        Gate::authorize('View Products List');
         $products = $this->productRepository->getAll();
         return $this->dataResponse($products);
     }
@@ -27,6 +29,7 @@ class ProductController extends Controller
      */
     public function store(ProductRequest $request)
     {
+        Gate::authorize('Create Product');
         $product = $this->productRepository->create($request->only(['sku', 'name', 'unit_price', 'stock']));
         return $this->successResponse('Exito al crear el producto', $product);
     }
@@ -36,6 +39,7 @@ class ProductController extends Controller
      */
     public function show(Request $request, string $id)
     {
+        Gate::authorize('View Product');
         $product = $this->productRepository->findById($id);
         return $this->dataResponse($product);
     }
@@ -43,8 +47,9 @@ class ProductController extends Controller
     /**
      * Update the specified resource in storage.
      */
-    public function update(Request $request, string $id)
+    public function update(ProductRequest $request, string $id)
     {
+        Gate::authorize('Update Product');
         $this->productRepository->update($id, $request->all());
         return $this->successResponse('Exito al actualizar el producto');
     }
@@ -54,6 +59,7 @@ class ProductController extends Controller
      */
     public function destroy(string $id)
     {
+        Gate::authorize('Delete Product');
         $this->productRepository->deleteById($id);
         return $this->successResponse('Exito al eliminar el producto');
     }
